Processing for the Advisor being able to click on a student now works.
There needs to be CSS added to the button for the students as it looks
bad at the moment.

Rewrote a bit of the User and Student classes to allow for students to
be created on login as well as a new student created at runtime by just
passing a eNumber to the function.

In the Advisor.jsp, all the rows are contained inside of a form for the
buttons.